chore(clippy): clean apohara-indexer + apohara-sandbox -D warnings debt
After autofix (commit 6e61506) the remaining manual fixes:

apohara-indexer/src/:
- db.rs/index.rs/indexer.rs: convert leading /// + blank-line module
  blocks to //! inner-doc comments (empty_line_after_doc_comments)
- parser.rs: #[allow(dead_code)] on parse_typescript_import_clause +
  find_import_source (partial implementations reserved for Stage 5+
  symbol resolution)
- parser.rs: collapse function_declaration / class_declaration branch
  into a single matches!() arm (if_same_then_else); both already produce
  ExportKind::Default(name)
- embeddings.rs: #[allow(clippy::large_enum_variant)] on EmbeddingModel;
  Real holds Bert + Tokenizer behind Arc already, boxing would just add
  indirection
- indexer.rs: add Indexer::is_empty() paired with public len();
  consume it in Indexer::search to silence len_zero
- indexer.rs: #[allow(dead_code)] on ModelLoadLock.0 (the File is held
  only to keep the flock alive — its data is never read)
- indexer.rs: add .truncate(false) to the model-lock OpenOptions
  (suspicious_open_options); concurrent flock holders must not see a
  zeroed file mid-flight
- index.rs: replace std::f32::{NAN,INFINITY} with f32::{NAN,INFINITY}
  (legacy_numeric_constants, x3)
- parser.rs: drop useless `assert!(imports.len() >= 0)` in the
  typescript-require test; the upstream unwrap already asserts the
  parser ran
- dependency.rs: .or_insert_with(Vec::new) → .or_default() (autofix
  carried over from Stage 6)

apohara-sandbox/src/profile/linux.rs:
- Remove redundant `req as u64` cast in the ioctl seccomp rule literal
  (unnecessary_cast); libc constants are already c_ulong on Linux.

Verification:
- cargo clippy --workspace --all-targets --no-deps -- -D warnings: clean
- cargo test -p apohara-indexer --lib: 67 passed
- cargo test -p apohara-indexer --test memory_integration: 9 passed
- cargo test -p apohara-indexer --test indexer_persistence: 2 passed
- cargo test -p apohara-sandbox: 2 e2e failures pre-existing
  (workspace_write_echo_succeeds, workspace_write_captures_stderr —
  futex/SIGSEGV in seccomp shell, unrelated to lint cleanup; reproduced
  with these changes stashed)

+++ crates/apohara-sandbox/src/profile/linux.rs
@@ -145,7 +145,7 @@ impl LinuxProfile {
                             "index": 1,
                             "type": "dword",
                             "op": "eq",
-                            "val": req as u64,
+                            "val": req,
                         }]
                     }));
                 }
